refactor: split query command into submodules
Split `cmd/query.rs` (1114 lines) into:
- `cmd/query/mod.rs` (230 lines) — Args, OutputFormat, run()
- `cmd/query/output.rs` (420 lines) — text/csv/json/beancount formatting
- `cmd/query/interactive.rs` (436 lines) — REPL loop and dot-commands

No behavioral changes. All files under 440 lines.
